ruff linting

+++ vllm/model_executor/layers/quantization/fp8.py
@@ -436,7 +436,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
 
         self.fused_experts: Optional[
             mk.FusedMoEModularKernel] = None  # type: ignore
-        
+
         # For GPUs that lack FP8 hardware support, we can leverage the Marlin
         # kernel for fast weight-only FP8 quantization
         self.use_marlin = (not current_platform.has_device_capability(89)
@@ -447,15 +447,14 @@ class Fp8MoEMethod(FusedMoEMethodBase):
 
         # First check for Flashinfer MOE on Blackwell GPUs
         self.flashinfer_moe_backend: Optional[FlashinferMoeBackend] = None
-        if (current_platform.is_cuda() and 
-            current_platform.is_device_capability(100) and
-            envs.VLLM_USE_FLASHINFER_MOE_FP8 and has_flashinfer_moe()):
-                self.flashinfer_moe_backend = get_flashinfer_moe_backend()
-                logger.info_once(
-                  f"Detected Blackwell GPUs, using FlashInfer "
-                  f"{self.flashinfer_moe_backend.value} kernels for FP8 MOE."
-                )
-        
+        if (current_platform.is_cuda()
+                and current_platform.is_device_capability(100)
+                and envs.VLLM_USE_FLASHINFER_MOE_FP8 and has_flashinfer_moe()):
+            self.flashinfer_moe_backend = get_flashinfer_moe_backend()
+            logger.info_once(
+                f"Detected Blackwell GPUs, using FlashInfer "
+                f"{self.flashinfer_moe_backend.value} kernels for FP8 MOE.")
+
         # Check for DeepGemm support.
         self.allow_deep_gemm = False
         if envs.VLLM_USE_DEEP_GEMM:
@@ -482,11 +481,10 @@ class Fp8MoEMethod(FusedMoEMethodBase):
         elif (current_platform.is_cuda()
               and current_platform.is_device_capability(100)
               and not self.flashinfer_moe_backend):
-                logger.info_once(
-                    "Using CutlassBlockScaledGroupedGemm kernels for Fp8 MOE "
-                    "on SM100."
-                )
-                self.allow_cutlass_block_scaled_grouped_gemm = True
+            logger.info_once(
+                "Using CutlassBlockScaledGroupedGemm kernels for Fp8 MOE "
+                "on SM100.")
+            self.allow_cutlass_block_scaled_grouped_gemm = True
 
     def create_weights(self, layer: Module, num_experts: int, hidden_size: int,
                        intermediate_size_per_partition: int,
@@ -941,8 +939,8 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                 import vllm.model_executor.layers.fused_moe.flashinfer_trtllm_moe  # noqa: E501, F401
                 assert (renormalize and use_grouped_topk
                         and custom_routing_function is None)
-                e_score_correction_bias = (e_score_correction_bias.to(x.dtype)
-                            if e_score_correction_bias is not None else None)
+                e_score_correction_bias = (e_score_correction_bias.to(
+                    x.dtype) if e_score_correction_bias is not None else None)
                 return torch.ops.vllm.flashinfer_fused_moe_blockscale_fp8(
                     routing_logits=router_logits.to(torch.float32),
                     routing_bias=e_score_correction_bias,
